pyproject.toml
[tool.poetry]
name = "API_bit.ly_Devman"
version = "0.1.0"
description = ""
authors = ["Roman Okatev <nebulawalker@live.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
requests = "^2.28.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
